feat: Add menu_handler to start_handler in handlers.py
Added menu_handler to start_handler in handlers.py to provide users with options to choose from after checking their subscription status. InlineKeyboardMarkup is used to create a menu with three buttons: "Catalog", "Cart", and "FAQ". If the user is subscribed, the menu is displayed, otherwise a message is sent saying "You are not subscribed to the channel!".