feat: Add CartProduct model for specifying the quantity of products in the cart
This commit adds a new model called CartProduct to the existing models. This model is used to specify the quantity of products in the cart. Previously, the quantity field in the Product model was not suitable for this purpose as it was not unique across different carts.

The CartProduct model has a foreign key to both the Cart and Product models and an additional field called quantity which specifies the quantity of a specific product in the cart. This model solves the issue of non-unique quantity field in the Product model and provides a better way to manage products in the cart.