feat: Added cart functionality
Added the ability for users to add products to their cart, confirm their purchase, and view their cart. The process_cart_add function was updated to import models related to the cart functionality and a new function cart_ask_confirmation_callback was added to ask the user to confirm their purchase. Two new functions were also added: add_product_to_cart to add the product to the user's cart and cart_show to display the cart.